% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pereira.R
\name{Mdelta}
\alias{Mdelta}
\title{Make delta}
\usage{
Mdelta(omega, lambda)
}
\arguments{
\item{omega}{dispersion parameter, stand for sigma square}

\item{lambda}{numeric, shape/skewness parameter}
}
\description{
Make delta
}
