% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pereira.R
\name{MOmegabar}
\alias{MOmegabar}
\title{Make scale matrix or scale scalar}
\usage{
MOmegabar(omega, delta, lambda)
}
\arguments{
\item{omega}{dispersion parameter, stand for sigma square}

\item{delta}{numeric, shape/skewness vector coming from \code{\link{Mdelta}}}

\item{lambda}{numeric, shape/skewness parameter}
}
\description{
Make scale matrix or scale scalar
}
