% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pereira.R
\name{nlmm.frame}
\alias{nlmm.frame}
\title{Frame for nlmm}
\usage{
nlmm.frame(
  formula,
  X.formula,
  fixef.name,
  cluster,
  family = "Gamma",
  data,
  subset,
  Zi = NULL,
  fixedStart,
  na.action = options("na.action")$na.action
)
}
\description{
Frame for nlmm
}
