% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normal.R
\name{Normal}
\alias{Normal}
\alias{dmvnorm}
\alias{dmnorm}
\alias{rmvnorm}
\title{Multivariate Normal Distribution}
\usage{
dmvnorm (x, mean = rep(0, p), Var = diag(p), log = FALSE)

dmnorm (x, mean = rep(0, p), Var = diag(p), log = FALSE)

rmvnorm (n, mean = rep(0, nrow(Var)), Var = diag(length(mean)))
}
\arguments{
\item{x}{vector or matrix of quantiles. If x is a matrix, each row is taken to be a quantile.}

\item{mean}{mean vector.}

\item{Var}{covariance matrix.}

\item{log}{logical; if \code{TRUE}, densities \code{d} are given as \code{log(d)}.}

\item{n}{number of deviates to be generated.}
}
\description{
Multivariate Normal Density function and Random deviate generator
}
\details{
See \code{heavy::rmnorm} for a faster alternative to \code{rmvnorm}, and \code{dmnorm} is just an alias for \code{dmvnorm}.
}
