% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pereira.R
\name{pereEM}
\alias{pereEM}
\title{Pereira EM algorithm (fit)}
\usage{
pereEM(
  formula,
  X.formula,
  fixef.name = fixef.name,
  Zi = NULL,
  cluster,
  family = "Gamma",
  data,
  subset,
  na.action = options("na.action")$na.action,
  start = list(alpha = c(100, 700, 349), D = diag(c(0.5, 1, 1.2)), lambda = rep(4, 3),
    sigma2 = 3, nu = NULL),
  pverbose = TRUE,
  maxit = 40,
  tol = 0.001,
  maxit.opt = 16,
  skew = TRUE
)
}
\description{
Pereira EM algorithm (fit)
}
