% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pereira.R
\name{sqrtm}
\alias{sqrtm}
\title{Matrix and numeric square root}
\usage{
sqrtm(x)
}
\arguments{
\item{x}{A matrix or numeric value.}
}
\value{
Square root of x.
}
\description{
Computes Matrix and numeric number square root using eigen analysis.
}
\examples{
sqrtm(4) # returns numeric
sqrtm(matrix(64)) # returns matrix
A <- matrix(data = c(1, 2, 3, 2, 20, 26, 3, 26, 70),
ncol = 3); A
a <- sqrtm(A); a
a \%*\% a ## ==A; matrix square
a * a ## not equals A; simple square
}
\seealso{
\code{\link{sqrt}}
}
