% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmssmn.R
\name{SSMNormal}
\alias{SSMNormal}
\alias{dmssmn}
\alias{dssmn}
\alias{rssmn}
\alias{rmssmn}
\title{(Multivariate) Skew Scale Mixture of Normal Distributions}
\usage{
dssmn (x, mu = 0, omegabar = 1, delta = 0, mixvar = "Gamma",
       df = switch(mixvar, Bin = c(.5, .5), 5), log = FALSE)

rssmn (n, mu = 0, omegabar = 1, delta = 0, mixvar = "Gamma",
       df = switch(mixvar, Bin = c(.5, .5), 5))

dmssmn (x, mu = rep(0, q), Omegabar = diag(q), delta = rep(0, q),
        mixvar = "Gamma", df = switch(mixvar, Bin = c(.5, .5), 5),
        log = FALSE)

rmssmn (n, mu = rep(0, q), Omegabar = diag(q), delta = rep(0, q),
        mixvar = "Gamma", df = switch(mixvar, Bin = c(.5, .5), 5))
}
\arguments{
\item{x}{vector or matrix of quantiles. If \code{x} is a matrix, each row is taken to be a quantile.}

\item{mu}{numeric, location vector.}

\item{Omegabar, omegabar}{numeric, scale matrix. \code{omegabar} is a scalar, for the univariate case.}

\item{delta}{numeric, shape/skewness vector.}

\item{mixvar}{character, indicates the distribution of the scale mixing variable of the target skew scale mixture of normal (ssmn) distribution. "Gamma" gives the (skew) t, "Beta" gives (skew) slash, "Bin" gives the (skew) contaminated normal, and any other string gives the (skew) normal.}

\item{df}{numeric, degrees of freedom parameter. It is ignored for (skew) normal distributions. Otherwise, \code{df} is a scalar for all non (skew) normal distributions, except, \code{df} is a vector of two elements each in the open (0, 1) when \code{mixvar = "Bin"}.
Conditions for the first order moment of the ssmn distribution to be defined: \code{df > 1} if \code{mixvar = "Gamma"}, \code{df > .5} if \code{mixvar = "Beta"}.
Conditions for the second order moment of the ssmn distribution to be defined: \code{df > 2} if \code{mixvar = "Gamma"}, \code{df > 1} if \code{mixvar = "Beta"}.}

\item{log}{logical; if \code{TRUE}, densities \code{d} are given as \code{log(d)}.}

\item{n}{integer, number of deviates to be generated.}
}
\description{
Random deviate generator for (Multivariate) Skew Scale Mixture of Normal Distributions
}
\details{
A ssmn variable/vector \code{X} is represented as:

\code{X = mu + T_0 * delta / sqrt(kappa) + z / sqrt(kappa)} where \code{T_0} is a half normal variable, \code{kappa} is a scale mixing variable, and \code{z} is a normal variable/vector with null mean and covariance matrix \code{Omegabar}.
}
